Split layout algorithm into tvOS and iOS
