+++ Sources/iOS+tvOS/Classes/FamilyScrollView.swift
@@ -29,11 +29,8 @@ public class FamilyScrollView: UIScrollView, FamilyDocumentViewDelegate, UIGestu
   /// See `observeView` methods for more information about which
   /// properties that get observed.
   private var observers = [Observer]()
-
-  private lazy var spaceManager = FamilySpaceManager()
-
+  internal lazy var spaceManager = FamilySpaceManager()
   lazy var cache = FamilyCache()
-
   private var isScrolling: Bool { return isTracking || isDragging || isDecelerating }
 
   /// The custom distance that the content view is inset from the safe area or scroll view edges.
@@ -208,7 +205,7 @@ public class FamilyScrollView: UIScrollView, FamilyDocumentViewDelegate, UIGestu
 
   /// Computes the content size for the collection view based on
   /// combined content size of all the underlaying scroll views.
-  private func computeContentSize() {
+  internal func computeContentSize() {
     let computedHeight = subviewsInLayoutOrder
       .filter({ $0.isHidden == false || ($0 as? FamilyWrapperView)?.view.isHidden == false })
       .reduce(CGFloat(0), { value, view in
@@ -285,124 +282,16 @@ public class FamilyScrollView: UIScrollView, FamilyDocumentViewDelegate, UIGestu
   /// based on the views index inside `subviewsInLayoutOrder`. This is invoked
   /// when a view changes size or origin. It also scales the frame of scroll views
   /// in order to keep dequeuing for table and collection views.
-  private func runLayoutSubviewsAlgorithm() {
-    if cache.isEmpty {
-      var yOffsetOfCurrentSubview: CGFloat = 0.0
-      for scrollView in subviewsInLayoutOrder where scrollView.isHidden == false {
-        let view = (scrollView as? FamilyWrapperView)?.view ?? scrollView
-        let insets = spaceManager.customInsets(for: view)
-        if (scrollView as? FamilyWrapperView)?.view.isHidden == true {
-          continue
-        }
 
-        var frame = scrollView.frame
-        var contentOffset = scrollView.contentOffset
+  @objc func injected() {
 
-        if self.contentOffset.y < yOffsetOfCurrentSubview {
-          contentOffset.y = insets.top
-          frame.origin.y = floor(yOffsetOfCurrentSubview)
-        } else {
-          contentOffset.y = self.contentOffset.y - yOffsetOfCurrentSubview
-          frame.origin.y = floor(self.contentOffset.y)
-        }
-
-        let remainingBoundsHeight = fmax(bounds.maxY - yOffsetOfCurrentSubview, 0.0)
-        let remainingContentHeight = fmax(scrollView.contentSize.height - contentOffset.y, 0.0)
-        var newHeight: CGFloat = ceil(fmin(remainingBoundsHeight, remainingContentHeight))
-
-        if scrollView is FamilyWrapperView {
-          newHeight = fmin(documentView.frame.height, scrollView.contentSize.height)
-        } else {
-          newHeight = fmin(documentView.frame.height, newHeight)
-        }
-
-        #if os(tvOS)
-        newHeight = fmin(documentView.frame.height, scrollView.contentSize.height)
-        #endif
-
-        let shouldScroll = (self.contentOffset.y >= frame.origin.y &&
-          self.contentOffset.y <= scrollView.frame.maxY) &&
-          frame.height >= documentView.frame.height
-
-        if shouldScroll {
-          scrollView.contentOffset.y = contentOffset.y
-        } else {
-          frame.origin.y = yOffsetOfCurrentSubview
-        }
-
-        frame.size.width = self.frame.size.width - insets.left - insets.right
-        frame.size.height = newHeight
-
-        if scrollView.frame != frame {
-          scrollView.frame = frame
-        }
-
-        yOffsetOfCurrentSubview += scrollView.contentSize.height + insets.bottom + insets.top
-        var cachedOrigin = frame.origin
-        cachedOrigin.y += insets.top
-        cache.add(entry: FamilyCacheEntry(view: view,
-                                          origin: cachedOrigin,
-                                          contentSize: scrollView.contentSize))
-      }
-      computeContentSize()
-    } else {
-      for scrollView in subviewsInLayoutOrder where scrollView.isHidden == false {
-        let view = (scrollView as? FamilyWrapperView)?.view ?? scrollView
-        guard let entry = cache.entry(for: view) else { continue }
-        if (scrollView as? FamilyWrapperView)?.view.isHidden == true {
-          continue
-        }
-
-        var frame = scrollView.frame
-        var contentOffset = scrollView.contentOffset
-
-        if self.contentOffset.y < entry.origin.y {
-          contentOffset.y = 0.0
-          frame.origin.y = floor(entry.origin.y)
-        } else {
-          contentOffset.y = self.contentOffset.y - entry.origin.y
-          frame.origin.y = floor(self.contentOffset.y)
-        }
-
-        let remainingBoundsHeight = fmax(bounds.maxY - entry.origin.y, 0.0)
-        let remainingContentHeight = fmax(entry.contentSize.height - contentOffset.y, 0.0)
-        var newHeight: CGFloat = ceil(fmin(remainingBoundsHeight, remainingContentHeight))
-
-        if scrollView is FamilyWrapperView {
-          newHeight = fmin(documentView.frame.height, scrollView.contentSize.height)
-        } else {
-          newHeight = fmin(documentView.frame.height, newHeight)
-        }
-
-        #if os(tvOS)
-          newHeight = fmin(documentView.frame.height, scrollView.contentSize.height)
-        #endif
-
-        let shouldScroll = (self.contentOffset.y > frame.origin.y &&
-          self.contentOffset.y < entry.maxY) &&
-          frame.height >= documentView.frame.height
-
-        if shouldScroll {
-          scrollView.contentOffset.y = contentOffset.y
-        } else {
-          frame.origin.y = entry.origin.y
-        }
-
-        frame.size.height = newHeight
-
-        if compare(scrollView.frame.origin, to: frame.origin) ||
-           compare(scrollView.frame.size, to: frame.size) {
-          scrollView.frame = frame
-        }
-      }
-    }
   }
 
-  private func compare(_ lhs: CGSize, to rhs: CGSize) -> Bool {
+  internal func compare(_ lhs: CGSize, to rhs: CGSize) -> Bool {
     return (abs(lhs.height - rhs.height) <= 0.001)
   }
 
-  private func compare(_ lhs: CGPoint, to rhs: CGPoint) -> Bool {
+  internal func compare(_ lhs: CGPoint, to rhs: CGPoint) -> Bool {
     return (abs(lhs.y - rhs.y) <= 0.001)
   }
 }
